Move geFormElementHelper into its own class
