+++ src/classes/FORM_PermissionValidator.cls
@@ -119,7 +119,6 @@ public with sharing class FORM_PermissionValidator {
                                     }
                                 }
                             }
-
                         }
                     }
                 }
