+++ src/lwc/geHelpers/geHelpers.js-meta.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<LightningComponentBundle xmlns="http://soap.sforce.com/2006/04/metadata">
+    <apiVersion>49.0</apiVersion>
+    <description>Ge Helpers</description>
+    <isExposed>false</isExposed>
+    <masterLabel>Ge Helpers</masterLabel>
+</LightningComponentBundle>
