+++ src/package.xml
@@ -3769,6 +3769,7 @@
         <members>geFormWidgetRowAllocation</members>
         <members>geFormWidgetTokenizeCard</members>
         <members>geGiftEntryFormApp</members>
+        <members>geHelpers</members>
         <members>geHome</members>
         <members>geLabelService</members>
         <members>geListView</members>
